Replace labs with annotations
It proved nearly impossible to position titles, subtitles, and captions
inside the plot area instead of having them on top and at the bottom
of the plot. Using annotations solves this but adds a ton of tuning
parameters. After the basemap has proved useful, they can be
moved to separate functions...

Caption is "label" instead of "text" like title and subtitle because the
caption is placed over the sea and we need a white background.