Update renv and packages (#61)
* Upgrade renv to 0.15.4 and update all packages

* Remove warning

Warning: Using `across()` in `filter()` is deprecated, use `if_any()` or `if_all()`.

* Rename column

readr package was upgraded. New version has different
naming convention for missing column names.

* Hide message about missing first column

readr apckage was upgraded and warning/message
convention changed

* Remove polygon borders from plots

These plots have a separate item in the legend for NA values.
It is not supported by ggplot2 natively for sf data. This is a hacky way
to achieve the wanted result. The previous method was also hacky,
but broke when ggplot2 was upgraded.