Decrease emissions due to speed limits
Helmet can predict emission reduction with mode shift and
destination-choice change, but not when the driving style is
more economical. This calculation is based on ELY report:
https://www.doria.fi/bitstream/handle/10024/182160/Raportteja%2057%202021.pdf
If speed limits are exactly the same as in the report, we aim for
2.6 % decrease (Helmet + economical driving style, page 37). If they
are lower (eg. to Kehä III) we assume, that it changes linearly.